offboarding users works when given an array of users works
